04-memory/requirements.txt
# 占位依赖（后续按需添加）
# 示例：langchain

# 核心
langchain>=1.2.0
langchain-core>=1.2.2

# OpenAI 提供商（使用 OpenAI API 时需要）
langchain-openai>=1.1.5

# 本地模型（使用 Ollama 时需要）
langchain-ollama>=1.0.1

langchain_community>=0.4.1
# 常用辅助
tiktoken>=0.12.0
python-dotenv>=1.2.1
